requirements.txt
opencv-python~=4.10.0.84
kagglehub~=0.2.7
torchvision~=0.19.0
tqdm~=4.66.4
pillow~=10.4.0
matplotlib~=3.8.4
seaborn~=0.13.2
numpy~=1.26.4
scikit-learn~=1.5.1